code update: check if corp members table exists in case we're reverting seeder on an empty db

+++ seeders/20210524122818-corp-member.js
@@ -51,15 +51,18 @@ module.exports = {
      */
 
     try {
-      await queryInterface.bulkDelete(
-        "CorpMembers",
-        {
-          id: {
-            [Op.in]: [1, 2],
+      const corp_members_table_exists = await queryInterface.tableExists('CorpMembers')
+      if (corp_members_table_exists) {
+        await queryInterface.bulkDelete(
+          "CorpMembers",
+          {
+            id: {
+              [Op.in]: [1, 2],
+            },
           },
-        },
-        {}
-      );
+          {}
+        );
+      }
     } catch (error) {
       console.error("Did not seed(down)", __filename, error);
     }
